test: segments user tests for better readability
